refactor: QuizUIState를 QuizViewModel 파일 안으로 이동
