+++ presentation/src/main/java/com/paykidscompose/presentation/model/QuizUIModel.kt
@@ -1,8 +1,6 @@
 package com.paykidscompose.presentation.model
 
-import com.paykidscompose.common.exception.PayKidsException
-import com.paykidscompose.presentation.base.UIState
-import com.paykidscompose.presentation.model.type.QuizType
+import com.paykidscompose.common.enums.QuizType
 
 data class QuizUIModel(
     val stage: Int,
@@ -14,14 +12,3 @@ data class QuizUIModel(
     val imageUrl: List<String>?,
     val totalCount: Int = 0
 ): UIModel()
-
-data class QuizUIState(
-    override val isLoading: Boolean = false,
-    override val error: PayKidsException? = null,
-    val quizzes: MutableList<QuizUIModel> = mutableListOf(),
-    val totalCount: Int = 0,
-    val currentIndex: Int = 0,
-): UIState() {
-    val currentQuiz: QuizUIModel?
-        get() = quizzes.getOrNull(currentIndex)
-}
